first version working with a hardcoded plugin operation
Probably not ready to ship in main yet, as it would be nice to have
something stronger. Also, faced SEGFAULT when giving a *mut
Box<struct> instead of a *mut Box<dyn PluginizableConnection> (as
it has some additional vtables there). Performance sadly decreased,
but there is no more generic types on most of the structures and
traits now.