A first working version of the Bytes API
In most cases, plugins should not have full direct access to raw
bytes that may be available in its context (for instance, a plugin
writing a new frame should not be able to read previous bytes), and
similarly we want to keep track on what bytes are read as well (for
instance, a plugin trying to read a frame should not be able to
read following ones without being spotted). We also need to keep
the overall overhead low.

This commit introduces the `Bytes` variant for `PluginVal` and two
new functions in the plugin API, `get_bytes()` and `put_bytes()`.
The current version works with copied `Vec<u8>`, but this design
should not precude working with zero-copy buffers such as `octets`
ones.