Store is per-plugin
The `Store` actually contains the memory offered to the plugins.
Given we want isolation between them, there is no point in keeping
a shared `Store` in the `PluginHandler`, and this can actually be
given to the `Plugin` structure, resulting in much less `unsafe`
code.

Furthermore, passing most of the "fake" immutable interactions with
the `PluginHandler` to mutable ones is actually quite easy to do,
and again remove a bunch of `unsafe` code. We might thus be on the
right design...