hide the UNIX-based `Instant` from the host
The previous version of pluginop forced the host to adopt using the
`unix_time::Instant` instead of the `std::time::Instant`. However,
this may not be convenient at all if the dependencies of the core
have strong dependencies on `std::time::Instant` (like Tokio or
future-based programming).

Instead, internally keep some reference times and perform the
conversion when needed.